refactor: reorganize directory structure
- hooks/ → composables/ (Vue convention for useXxx)
- plugin.ts → plugin/install.ts (group plugin internals)
- mixin.ts → plugin/lifecycle-tracker.ts (clearer intent)
- Rename createVRTMixin → createLifecycleTracker
