+++ src/composables/useRenderDiagnostics.ts
