+++ src/index.ts
@@ -1,6 +1,6 @@
-export { VueRenderDiagnostics } from './plugin.ts';
-export { useRenderDiagnostics } from './hooks/useRenderDiagnostics.ts';
-export type { UseRenderDiagnosticsReturn } from './hooks/useRenderDiagnostics.ts';
+export { VueRenderDiagnostics } from './plugin/install.ts';
+export { useRenderDiagnostics } from './composables/useRenderDiagnostics.ts';
+export type { UseRenderDiagnosticsReturn } from './composables/useRenderDiagnostics.ts';
 export type {
   VRTComponentLog,
   VRTIssue,
