+++ src/plugin/install.ts
@@ -1,8 +1,8 @@
 import type { Plugin } from 'vue';
-import type { VRTPluginOptions } from './types.ts';
-import { VRT_COLLECTOR_KEY } from './constants.ts';
-import { Collector } from './core/collector.ts';
-import { createVRTMixin } from './hooks/mixin.ts';
+import type { VRTPluginOptions } from '../types.ts';
+import { VRT_COLLECTOR_KEY } from '../constants.ts';
+import { Collector } from '../core/collector.ts';
+import { createLifecycleTracker } from './lifecycle-tracker.ts';
 
 export const VueRenderDiagnostics: Plugin<[VRTPluginOptions?]> = {
   install(app, options: VRTPluginOptions = {}) {
@@ -10,6 +10,6 @@ export const VueRenderDiagnostics: Plugin<[VRTPluginOptions?]> = {
 
     const collector = new Collector(options.thresholds);
     app.provide(VRT_COLLECTOR_KEY, collector);
-    app.mixin(createVRTMixin(collector, options));
+    app.mixin(createLifecycleTracker(collector, options));
   },
 };
